move ConfigDialog to common package.
